applied DTO's to adminController

+++ src/main/java/com/journalApp/service/UserService.java
@@ -1,18 +1,23 @@
 package com.journalApp.service;
 
 import com.journalApp.dto.CreateUserDto;
+import com.journalApp.dto.UpdateUserDto;
 import com.journalApp.entity.User;
 import com.journalApp.repository.UserRepository;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.coyote.BadRequestException;
+import org.apache.kafka.common.errors.DuplicateResourceException;
 import org.apache.kafka.common.errors.ResourceNotFoundException;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.mongodb.core.schema.MongoJsonSchema;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.client.HttpClientErrorException;
 
 import java.util.Arrays;
 import java.util.List;
@@ -66,14 +71,22 @@ public class UserService {
         userRepository.save(user);
     }
 
-    public ResponseEntity<String> saveAdmin(User user){
-        user.setPassword(passwordEncoder.encode(user.getPassword()));
-        user.setRoles(Arrays.asList("USER","ADMIN"));
-        userRepository.save(user);
-        return new ResponseEntity<>("Admin created \n"+
-                "userId :"+user.getId()+"\n"+
-                "UserName :"+user.getUsername()+"\n"
-                , HttpStatus.CREATED);
+    public User saveAdmin(CreateUserDto user) {
+        try {
+            if (userRepository.existsByUsername(user.getUsername()) || userRepository.existsByEmail(user.getEmail())) {
+                return null;
+            }
+            User inputUser = new User();
+            inputUser.setUsername(user.getUsername());
+            inputUser.setPassword(passwordEncoder.encode(user.getPassword()));
+            inputUser.setEmail(user.getEmail());
+            inputUser.setRoles(Arrays.asList("USER", "ADMIN"));
+            inputUser.setSentimentAnalysis(false);
+            userRepository.save(inputUser);
+            return inputUser;
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
     }
 
 
@@ -87,28 +100,42 @@ public class UserService {
     }
 
     public void deleteById(ObjectId id){
-         userRepository.deleteById(id);
+         User user= userRepository.findById(id).orElseThrow(() ->new ResourceNotFoundException("User not found with id "+id));
+         if(user!=null){
+             userRepository.deleteById(id);
+         }
     }
 
     public User findByusername(String username){
         return userRepository.findByUsername(username);
     }
 
-    @Transactional
-    public User updateUser(String id, User body) {
+    public User updateUser(String id, UpdateUserDto body){
         ObjectId objectId=new ObjectId(id);
         User existingUser=userRepository.findById(objectId).orElseThrow(()->new ResourceNotFoundException("User not found"));
 
         if(body.getUsername()!=null){
-            existingUser.setUsername(body.getUsername());
+           User existed= userRepository.findByUsername(body.getUsername());
+           if(existed==null) existingUser.setUsername(body.getUsername());
+           else throw new DuplicateResourceException("username already exists");
         }
+
+        if(body.getEmail()!=null) {
+            if (body.getEmail().equals(existingUser.getEmail()) || userRepository.existsByEmail(body.getEmail())) {
+                throw new DuplicateResourceException("email already exists");
+            }
+            if (!body.getEmail().equals(existingUser.getEmail())) {
+                emailService.sendEmail(existingUser.getEmail(), "Changing Email 🔗", "Hi " + existingUser.getUsername() + " we are informing that you have changed your email,\n From now all the new emails are sent to your new email " + body.getEmail() + ".\n" + "If its not u then click on the link below , thank you.");
+                emailService.sendEmail(body.getEmail(), "Welcome to JournalApp " + body.getUsername() + " 🚀", "You have successfully registered to journalApp, now you can create journals for your daily schedule to improve yourself. \n Get the best of you✅");
+                existingUser.setEmail(body.getEmail());
+            }
+        }
+
         if(body.getPassword()!=null){
-            existingUser.setPassword(body.getPassword());
+            existingUser.setPassword(passwordEncoder.encode(body.getPassword()));
         }
-        if(body.getEmail()!=null){
-            emailService.sendEmail(existingUser.getEmail(),"Changing Email 🔗","Hi "+existingUser.getUsername()+" we are informing that you have changed your email,\n From now all the new emails are sent to your new email "+body.getEmail()+".\n"+"If its not u then click on the link below , thank you.");
-            emailService.sendEmail(body.getEmail(), "Welcome to JournalApp " + body.getUsername() + " 🚀", "You have successfully registered to journalApp, now you can create journals for your daily schedule to improve yourself. \n Get the best of you✅");
-            existingUser.setEmail(body.getEmail());
+        if(body.getSentimentAnalysis()!=null){
+            existingUser.setSentimentAnalysis(body.getSentimentAnalysis());
         }
        return userRepository.save(existingUser);
     }
